vessel-enhancement/requirements.txt
colorlog
cupy_cuda12x
dask
imageio
matplotlib
monai
nibabel
numpy
pytest
PyYAML
PyYAML
scikit_learn
scipy
simpleitk
skimage
tabulate
tifffile
torch
tqdm
